# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Contact API with Authentication, Roles, and Permissions"
  },
  "basePath": "/api",
...
  "securityDefinitions": {
    "cookieAuth": {
      "type": "apiKey",
      "name": "session",
      "in": "cookie",
      "description": "Session cookie obtained after successful login. Handled automatically by the browser."
    }
  },
  "paths": {
...
    "/app/categories/{category_id}": {
      "get": {
        "summary": "Get a single permission category by ID.",
        "tags": ["Category Management"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "The ID of the category to retrieve."
          },
          {
            "name": "include_usage",
            "in": "query",
            "type": "boolean",
            "description": "If true, includes usage statistics (number of associated permissions).",
            "default": false
          },
          {
            "name": "include_affected_permissions",
            "in": "query",
            "type": "boolean",
            "description": "If true, includes a list of affected permissions (ID, name, status). Requires `include_usage=true` to be effective.",
            "default": false
          }
        ],
        "responses": {
          "200": {
            "description": "Category details.",
            "schema": { "$ref": "#/definitions/Category" }
          },
          "401": { "$ref": "#/responses/UnauthorizedError" },
          "403": { "$ref": "#/responses/ForbiddenError" },
          "404": { "$ref": "#/responses/NotFoundError" }
        }
      },
      "patch": {
        "summary": "Update an existing permission category.",
        "tags": ["Category Management"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "The ID of the category to update."
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/CategoryUpdateInput" }
          }
        ],
        "responses": {
          "200": {
            "description": "Category updated successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "category": { "$ref": "#/definitions/Category" }
              }
            }
          },
          "400": { "$ref": "#/responses/BadRequestError" },
          "401": { "$ref": "#/responses/UnauthorizedError" },
          "403": { "$ref": "#/responses/ForbiddenError" },
          "404": { "$ref": "#/responses/NotFoundError" },
          "409": { "$ref": "#/responses/ConflictError" }
        }
      },
      "delete": {
        "summary": "Delete a permission category.",
        "tags": ["Category Management"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "The ID of the category to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully.",
            "schema": { "$ref": "#/definitions/SuccessMessage" }
          },
          "401": { "$ref": "#/responses/UnauthorizedError" },
          "403": { "$ref": "#/responses/ForbiddenError" },
          "404": { "$ref": "#/responses/NotFoundError" },
          "409": { "$ref": "#/responses/ConflictError" }
        }
      }
    }
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "description": { "type": "string", "x-nullable": true },
        "status": { "type": "string", "enum": ["active", "inactive"], "description": "The active status of the category.", "default": "active" },
        "usage": { "type": "integer", "description": "The number of permissions associated with this category. Only included if `include_usage=true`." },
        "affected_permissions": {
          "type": "array",
          "description": "A list of permissions (ID, name, status) associated with this category. Only included if `include_affected_permissions=true`.",
          "items": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" },
              "status": { "type": "string" }
            }
          }
        }
      }
    },
...
    "CategoryUpdateInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "The new unique name of the category."
        },
        "description": {
          "type": "string",
          "description": "An updated description of the category.",
          "x-nullable": true
        },
        "status": {
          "type": "string",
          "enum": ["active", "inactive"],
          "description": "The updated status of the category."
        }
      }
    },
...
    "SuccessMessage": {
      "type": "object",
      "properties": {
        "message": { "type": "string" }
      }
    },
    "ErrorMessage": {
      "type": "object",
      "properties": {
        "message": { "type": "string" }
      }
    },
...
  },
  "responses": {
    "UnauthorizedError": {
      "description": "Authentication information is missing or invalid. User is not logged in.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    },
    "ForbiddenError": {
      "description": "The user does not have the necessary permissions for a resource.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    },
    "NotFoundError": {
      "description": "The requested resource was not found.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    },
    "BadRequestError": {
      "description": "Invalid input provided.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    },
    "ConflictError": {
      "description": "Resource already exists (e.g., username, email, role name, permission name) or a dependency conflict.",
      "schema": { "$ref": "#/definitions/ErrorMessage" }
    }
  }
}